게시글 삭제 confirm 추가 & boardUpdate.css 수정 & miniHome.css !important 최소화

+++ FeatherWorld/src/main/resources/static/js/board/boardDetail.js
@@ -66,6 +66,10 @@ document.addEventListener("DOMContentLoaded", () => {
   if(boardDeleteBtn) {
     boardDeleteBtn.addEventListener("click", async () => {
 
+      if(!confirm("게시글을 정말 삭제하시겠습니까?")) {
+        return;
+      }
+      
       // 현재 게시글 번호(boardNo)로 게시글 삭제 요청
       const resp = await fetch(`/${memberNo}/board/${boardCode}/${boardNo}/delete`, {method: "delete"});
       const result = await resp.text();
